+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 2.0.3 (2017.06.12)
 + **cloudcomb:** add IP api
+* **cloudcomb:** Update Microservice api
 
 ## Version 2.0.2 (2017.05.19)
 + **cloudcomb:** add microservice api